Added ownership check for comment management

+++ Systems/API/ZooExpoOrg.Api/Controllers/Comments/CommentController.cs
@@ -6,6 +6,7 @@ using ZooExpoOrg.Common.Exceptions;
 using ZooExpoOrg.Common.Security;
 using ZooExpoOrg.Services.Comments;
 using ZooExpoOrg.Services.Logger;
+using ZooExpoOrg.Services.RightVerifier;
 
 namespace ZooExpoOrg.Api.Controllers.Comments;
 
@@ -18,16 +19,19 @@ public class CommentController : Controller
     private readonly IAppLogger logger;
     private readonly ICommentService commentService;
     private readonly IMapper mapper;
+    private readonly IRightVerifierService rightVerifier;
 
     public CommentController(
         IAppLogger logger, 
         ICommentService commentService, 
-        IMapper mapper
+        IMapper mapper,
+        IRightVerifierService rightVerifier
         )
     {
         this.logger = logger;
         this.commentService = commentService;
         this.mapper = mapper;
+        this.rightVerifier = rightVerifier;
     }
 
     [HttpGet("location/{locationId:Guid}")]
@@ -66,11 +70,18 @@ public class CommentController : Controller
 
     [HttpPost]
     [Authorize(AppScopes.UseScope)]
-    public async Task<IActionResult> Create(CreateCommentModel model)
+    public async Task<IActionResult> Create(PresintationCreateCommentModel model)
     {
         try
         {
-            var result = await commentService.Create(model);
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfCreateComment(jwtToken, model.AuthorId)))
+            {
+                return BadRequest("Access denied.");
+            }
+
+            var result = await commentService.Create(mapper.Map<CreateCommentModel>(model));
 
             return Ok(mapper.Map<PresintationCommentModel>(result));
         }
@@ -82,11 +93,18 @@ public class CommentController : Controller
 
     [HttpPut("{id:Guid}")]
     [Authorize(AppScopes.UseScope)]
-    public async Task<IActionResult> Update(Guid id, UpdateCommentModel model)
+    public async Task<IActionResult> Update(Guid id, PresintationUpdateCommentModel model)
     {
         try
         {
-            await commentService.Update(id, model);
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagComment(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
+            await commentService.Update(id, mapper.Map<UpdateCommentModel>(model));
 
             return Ok();
         }
@@ -102,6 +120,13 @@ public class CommentController : Controller
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagComment(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await commentService.Delete(id);
 
             return Ok();
